tags: Add image refcounting, so we can cleanup orphans
Before this commit there was no way to remove images that were no
longer needed by any tags.  This commit adds a persistent list of
references for each image, with both the depending image id and the
namespace, repository, and tag.  When we delete a tag, we traverse
it's ancestors removing any obsolete references and any images that
are no longer referenced.

This is more complicated than just having a refcount number, but also
more resilient.  For example, someone will eventually have a power
outage after removing a tag while the refcounts were still being
adjusted.  By keeping the references themselves in a list, it's easy
to check for stale references and remove them.  If it turns out that
the removal checks are too slow, it would be easy to make them more
optimistic (e.g. passing the descendant image and tags that we expect
to remove down the _remove_references stack and removing them instead
of checking all listed references for all ancestors).

The JSON serialization for the [namespace, repository, tag] key allows
us to use JSON to serialize the resulting dictionary (JSON only
supports string keys).  It's a bit awkward, but I wanted to make sure
a single tag was only represented once (with whatever its current
image was) in the references list.