test: Expect a 504 when the auth service times out
I've lowered the proxy timeout so we don't have to wait the default 60
seconds [1] for this test to complete.  I'm currently looking for a
way to pass the 504 from the auth service on to the client.  From the
current error logs, we can see that it's being translated to a 500
error:

  nginx_1 | 2014/10/23 20:40:56 [error] 6#0: *45 auth request unexpected status: 504 while sending to client, ...
  nginx_1 | 10.0.42.1 - - [23/Oct/2014:20:40:56 +0000] "GET / HTTP/1.1" 500 192 "-" "Python-urllib/3.4"

[1]: http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_read_timeout