proxy.conf: Mark the /auth endpoint as internal
There's no need to hide this away.  It's an auth endpoint, so in
production you'll serve it behind HTTPS, and the only responses you'll
get are:

* You need credentials for that endpoint, and you didn't give me any
* You need credentials for that endpoint, and yours are authorized
* You need credentials for that endpoint, and yours are not authorized
* Yes, you can use that endpoint

Still, the auth service is not something that the *registry* offers
(it's a separate service), so we shouldn't expose it directly through
this registry proxy.